feat: refactor loading component to tailwindcss styles and add new spinner component
