+++ src/pages/Pokemon/Pokemon.tsx
@@ -1,16 +1,16 @@
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import { Loading } from "../../components/Loading/Loading";
-import UniquePokemonPage from "../../components/UniquePokemonPage/UniquePokemonPage";
-import { usePokedex } from "../../hooks/usePokedex";
-import theme from "../../styles/theme";
-import useEvolution from "../../hooks/useEvolution";
-import { useForms } from "../../hooks/useForms";
+import { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
+import Loading from '../../components/Loading/Loading';
+import UniquePokemonPage from '../../components/UniquePokemonPage/UniquePokemonPage';
+import { usePokedex } from '../../hooks/usePokedex';
+import theme from '../../styles/theme';
+import useEvolution from '../../hooks/useEvolution';
+import { useForms } from '../../hooks/useForms';
 
 export function Pokemon() {
   const { getPokemonData, uniquePokemonData } = usePokedex();
   const { firstEvolution, secondEvolution, thirdEvolution } = useEvolution();
-  const {forms} = useForms()
+  const { forms } = useForms();
   const [isLoading, setIsLoading] = useState(true);
   const params = useParams();
 
@@ -21,7 +21,7 @@ export function Pokemon() {
   }, [params]);
 
   if (isLoading) {
-    return <Loading color={theme.colors.primary[500]} size={64} />;
+    return <Loading color={theme.colors.primary[500]} size={'lg'} />;
   } else {
     // console.log(uniquePokemonData);
     return (
